PhotonPropagation

Documentation for PhotonPropagation.

PhotonPropagation.Detection.Circular

PhotonPropagation.Detection.Rectangular

PhotonPropagation.Detection.Spherical

PhotonPropagation.Detection.TargetShape

PhotonPropagation.LightYield.CherenkovTrackLengthParameters

PhotonPropagation.LightYield.LightyieldParametrisation

PhotonPropagation.LightYield.LongitudinalParameterisation

PhotonPropagation.LightYield.LongitudinalParameterisation

PhotonPropagation.LightYield.LongitudinalParametersBase

PhotonPropagation.Medium.DIPPR105Params

PhotonPropagation.Medium.WaterProperties

PhotonPropagation.Medium.WaterProperties

PhotonPropagation.Detection.surface_area

PhotonPropagation.LightYield._interp_ch_ang_dist_int

PhotonPropagation.LightYield.cascade_cherenkov_track_length

PhotonPropagation.LightYield.cherenkov_ang_dist

PhotonPropagation.LightYield.cherenkov_ang_dist_int_analytic

PhotonPropagation.LightYield.gamma_cdf

PhotonPropagation.LightYield.get_longitudinal_params

PhotonPropagation.LightYield.integral_long_profile

PhotonPropagation.LightYield.longitudinal_profile

PhotonPropagation.LightYield.oversample_source

PhotonPropagation.LightYield.rel_additional_track_length

PhotonPropagation.LightYield.rel_additional_track_length_params

PhotonPropagation.Medium.DIPPR105

PhotonPropagation.Medium._absorption_length_straw

PhotonPropagation.Medium._calc_quan_fry_params

PhotonPropagation.Medium._sca_len_part_conc

PhotonPropagation.Medium.absorption_length

PhotonPropagation.Medium.c_at_wl

PhotonPropagation.Medium.cherenkov_angle

PhotonPropagation.Medium.dispersion

PhotonPropagation.Medium.dispersion_fry

PhotonPropagation.Medium.group_velocity

PhotonPropagation.Medium.make_cascadia_medium_properties

PhotonPropagation.Medium.refractive_index

PhotonPropagation.Medium.refractive_index_fry

PhotonPropagation.Medium.scattering_length

PhotonPropagation.PhotonPropagationCuda.cuda_hg_scattering_func

PhotonPropagation.PhotonPropagationCuda.initialize_direction_isotropic

PhotonPropagation.PhotonPropagationCuda.uniform

PhotonPropagation.Spectral.frank_tamm

PhotonPropagation.Spectral.frank_tamm_inverted_cdf

PhotonPropagation.Spectral.frank_tamm_norm

PhotonPropagation.Medium.DIPPR105Params — Type

DIPPR105Params

Parameters for the DIPPR105 formula

PhotonPropagation.Medium.WaterProperties — Type

WaterProperties{T<:Real} <: MediumProperties{T}

Properties for a water-like medium. Use unitful constructor to create a value of this type.

Fields:

-salinity – Salinity (permille) -pressure – Pressure (atm) -temperature – Temperature (°C) -volconcsmallpart – Volumetric concentrations of small particles (ppm) -volconclargepart – Volumetric concentrations of large particles (ppm) -radiationlength – Radiation length (g/cm^2) -density – Density (kg/m^3) -meanscattering_angle – Cosine of the mean scattering angle

PhotonPropagation.Medium.WaterProperties — Method

    function WaterProperties(
        salinity::Unitful.Quantity{T},
        pressure::Unitful.Quantity{T},
        temperature::Unitful.Quantity{T},
        vol_conc_small_part::Unitful.Quantity{T},
        vol_conc_large_part::Unitful.Quantity{T},
        radiation_length::Unitful.Quantity{T}
    ) where {T<:Real}
Construct a `WaterProperties` type.

The constructor uses DIPPR105 to calculate the density at the given temperature.
Parameters for the Quan-Fry parametrisation of the refractive index are calculated
for the given salinity, temperature and pressure.

PhotonPropagation.Medium.DIPPR105 — Function

DIPPR105(temperature::Real, params::DIPPR105Params=DDBDIPR105Params)

Use DPPIR105 formula to calculate water density as function of temperature. temperature in K.

Reference: http://ddbonline.ddbst.de/DIPPR105DensityCalculation/DIPPR105CalculationCGI.exe?component=Water

Returns density in kg/m^3

PhotonPropagation.Medium._absorption_length_straw — Method

_absorption_length_straw(wavelength::Real)

Calculate the absorption length at wavelength (in nm).

Based on interpolation of STRAW attenuation length measurement.

PhotonPropagation.Medium._calc_quan_fry_params — Method

_calc_quan_fry_params(salinity::Real, temperature::Real, pressure::Real)

Helper function to get the parameters for the Quan & Fry formula as function of salinity, temperature and pressure.

PhotonPropagation.Medium._sca_len_part_conc — Method

_sca_len_part_conc(wavelength; vol_conc_small_part, vol_conc_large_part)

Calculates the scattering length (in m) for a given wavelength based on concentrations of small (vol_conc_small_part) and large (vol_conc_large_part) particles. wavelength is given in nm, volconcsmallpart and volconclargepart in ppm

Adapted from clsim ©Claudio Kopper

PhotonPropagation.Medium.absorption_length — Method

absorption_length(wavelength, medium::MediumProperties)

Return absorption length (in m) at wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.c_at_wl — Method

c_at_wl(wavelength, medium)

Return speed of light in medium (m/ns) at wavelength wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.cherenkov_angle — Method

cherenkov_angle(wavelength, medium::MediumProperties)

Calculate the cherenkov angle (in rad) for wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.dispersion — Method

refractive_index(wavelength, medium)

Return the dispersion at wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.dispersion_fry — Method

dispersion_fry( wavelength::T; salinity::Real, temperature::Real, pressure::Real) where {T <: Real}

Calculate the dispersion (dn/dλ) for the Quan & Fry model.
wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

PhotonPropagation.Medium.group_velocity — Method

refractive_index(wavelength, medium)

Return the group_velocity in m/ns at wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.make_cascadia_medium_properties — Method

make_cascadia_medium_properties(::Type{T}) where {T<:Real}

Construct WaterProperties with properties from Cascadia Basin of numerical type T.

PhotonPropagation.Medium.refractive_index — Method

refractive_index(wavelength, medium)

Return the refractive index at wavelength.

wavelength is expected to be in units nm.

PhotonPropagation.Medium.refractive_index_fry — Method

refractive_index_fry(wavelength, salinity, temperature, pressure)

The phase refractive index of sea water according to a model from Quan & Fry.

wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

The original model is taken from: X. Quan, E.S. Fry, Appl. Opt., 34, 18 (1995) 3477-3480.

An additional term describing pressure dependence was included according to: Wolfgang H.W.A. Schuster, "Measurement of the Optical Properties of the Deep Mediterranean - the ANTARES Detector Medium.", PhD thesis (2002), St. Catherine's College, Oxford downloaded Jan 2011 from: http://www.physics.ox.ac.uk/Users/schuster/thesis0098mmjhuyynh/thesis.ps

Adapted from clsim (©Claudio Kopper)

PhotonPropagation.Medium.scattering_length — Method

scattering_length(wavelength, medium::MediumProperties)

Return scattering length at wavelength.

wavelength is expected to be in units nm. Returned length is in units m.

PhotonPropagation.Spectral.frank_tamm — Method

frank_tamm(wavelength::Real, ref_index::T) where {T<:Real}

Evaluate Frank-Tamm formula

PhotonPropagation.Spectral.frank_tamm_inverted_cdf — Method

frank_tamm_inverted_cdf(wl_range, step_size)

Return the inverted CDF of the Frank-Tamm Spectrum in range wl_range evaluated with step size step_size.

PhotonPropagation.Spectral.frank_tamm_norm — Method

frank_tamm_norm(wl_range::Tuple{T, T}, ref_index_func::Function) where {T<:Real}

Calculate number of Cherenkov photons per length in interval wl_range. Returned number is in units m^-1.

PhotonPropagation.Detection.Circular — Type

Circular{T <: Real} <: TargetShape

Circular target shape. CAUTION: The orientation is currently always e_z

Fields

position - center point of the rectangle (in m)

radius - circle radius (in m)

PhotonPropagation.Detection.Rectangular — Type

Rectangular{T <: Real} <: TargetShape

Rectangular target shape. CAUTION: The orientation is currently always e_z

Fields

position - center point of the rectangle (in m)

length_x - length in x direction (in m)

length_y - length in y direction (in m)

PhotonPropagation.Detection.Spherical — Type

Spherical{T <: Real} <: TargetShape

Spherical target shape

Fields

position - center point of the sphere (in m)

radius - sphere radius (in m)

PhotonPropagation.Detection.TargetShape — Type

TargetShape

Base type for photon target shapes.

PhotonPropagation.Detection.surface_area — Method

surface_area(::TargetShape) Return the surface area of the target shape.

PhotonPropagation.PhotonPropagationCuda.cuda_hg_scattering_func — Method

cuda_hg_scattering_func(g::Real)

CUDA-optimized version of Henyey-Greenstein scattering in one plane.

Arguments

g::Real: mean scattering angle

Returns

typeof(g) cosine of a scattering angle sampled from the distribution

PhotonPropagation.PhotonPropagationCuda.initialize_direction_isotropic — Method

initialize_direction_isotropic(T::Type)

Sample a direction isotropically

Arguments

T::Type: desired eltype of the return value

Returns

StaticVector{3, T}: Cartesian coordinates of the sampled direction

PhotonPropagation.PhotonPropagationCuda.uniform — Method

uniform(minval::T, maxval::T) where {T}

Convenience function for sampling uniform values in [minval, maxval]

PhotonPropagation.LightYield.CherenkovTrackLengthParameters — Type

Struct for storing the parameters of the Cherenkov track length parametrisation. See: https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield.LightyieldParametrisation — Type

Struct for holding both the longitudinal and the track length parametrisation

PhotonPropagation.LightYield.LongitudinalParameterisation — Type

Struct for storing the energy-corrected values of the longitudinal Cherenkov lightyield parametrisation.

PhotonPropagation.LightYield.LongitudinalParameterisation — Method

LongitudinalParameterisation(energy, medium::MediumProperties, ::Type{ptype})

Construct LongitudinalParameterisation for a particle of type ptype at energy energy

PhotonPropagation.LightYield.LongitudinalParametersBase — Type

LongitudinalParametersBase

Struct for storing the parameters of the longitudinal Cherenkov lightyield parametrisation. These parameters are the baseline values at E=1GeV.

For obtaining the energy-correcected parameters, see get_longitudinal_params.

See: https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield._interp_ch_ang_dist_int — Function

_interp_ch_ang_dist_int(ref_ix_low, ref_ix_hi, steps=0.01)

Interpolate the integral of the Cherenkov angular distance distribution in the range [refixlow, refixhi].

PhotonPropagation.LightYield.cascade_cherenkov_track_length — Method

cascade_cherenkov_track_length(energy::Real, track_len_params::CherenkovTrackLengthParameters)

Cherenkov track length for cascades. energy in GeV

returns track length in m

PhotonPropagation.LightYield.cherenkov_ang_dist — Function

cherenkovangdist(costheta, ref_index)

Angular distribution of cherenkov photons for EM cascades.

Taken from https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield.cherenkov_ang_dist_int_analytic — Function

cherenkov_ang_dist_int_analytic(ref_index, lower, upper, ang_dist_pars)

Integral of the cherenkov angular distribution function.

PhotonPropagation.LightYield.gamma_cdf — Method

gamma_cdf(a, b, z)

Cumulative Gamma distribution $ int_0^z Gamma(a, b) $

PhotonPropagation.LightYield.get_longitudinal_params — Method

get_longitudinal_params(::Type{ptype}, energy)

Calculate the longitudinal parameters at energy (in GeV).

PhotonPropagation.LightYield.integral_long_profile — Method

integral_long_profile(z_low::Real, z_high::Real, long_param::LongitudinalParameterisation)

Integral of the longitudinal profile from shower depth zlow (in m) to zhigh

PhotonPropagation.LightYield.longitudinal_profile — Method

longitudinal_profile(z::Real, medium::MediumProperties, long_param::LongitudinalParameterisation)

Longitudinal shower profile (PDF) at shower depth z ( in m )

PhotonPropagation.LightYield.oversample_source — Method

oversample_source(em::T, factor::Number) where {T <: PhotonSource}

Return a new emitter of type T with number of photons increased by factor

PhotonPropagation.LightYield.rel_additional_track_length — Method

rel_additional_track_length(ref_index)

Calculate additional track length for muons. From https://arxiv.org/pdf/1206.5530.pdf

PhotonPropagation.LightYield.rel_additional_track_length_params — Method

rel_additional_track_length_params(ref_index)

Interpolate additional track length parameters, taken from https://arxiv.org/pdf/1206.5530.pdf using E_cut = 500MeV