PhotonPropagation

Documentation for PhotonPropagation.

PhotonPropagation.Detection.Circular

PhotonPropagation.Detection.DummyShape

PhotonPropagation.Detection.HomogeneousDetector

PhotonPropagation.Detection.POMRelativeAcceptance

PhotonPropagation.Detection.Rectangular

PhotonPropagation.Detection.Spherical

PhotonPropagation.Detection.SphericalHomogeneousDetector

PhotonPropagation.Detection.SphericalMultiPMTDetector

PhotonPropagation.Detection.TargetShape

PhotonPropagation.LightYield.AxiconeEmitter

PhotonPropagation.LightYield.CherenkovTrackLengthParameters

PhotonPropagation.LightYield.CollimatedIsotropicEmitter

PhotonPropagation.LightYield.ExtendedCherenkovEmitter

PhotonPropagation.LightYield.ExtendedCherenkovEmitter

PhotonPropagation.LightYield.ExtendedCherenkovEmitterCustomSmear

PhotonPropagation.LightYield.ExtendedCherenkovEmitterCustomSmear

PhotonPropagation.LightYield.ExtendedCherenkovEmitterNoSmear

PhotonPropagation.LightYield.ExtendedCherenkovEmitterNoSmear

PhotonPropagation.LightYield.LightyieldParametrisation

PhotonPropagation.LightYield.LongitudinalParameterisation

PhotonPropagation.LightYield.LongitudinalParameterisation

PhotonPropagation.LightYield.LongitudinalParametersBase

PhotonPropagation.LightYield.PencilEmitter

PhotonPropagation.LightYield.PointlikeIsotropicEmitter

PhotonPropagation.LightYield.PointlikeTimeRangeEmitter

PhotonPropagation.Medium.HomogenousIceProperties

PhotonPropagation.Detection.check_pmt_hit

PhotonPropagation.Detection.check_pmt_hit

PhotonPropagation.Detection.check_pmt_hit_opening_angle

PhotonPropagation.Detection.get_pmt_positions

PhotonPropagation.Detection.surface_area

PhotonPropagation.LightYield._interp_ch_ang_dist_int

PhotonPropagation.LightYield.cascade_cherenkov_track_length

PhotonPropagation.LightYield.cherenkov_ang_dist

PhotonPropagation.LightYield.cherenkov_ang_dist_int_analytic

PhotonPropagation.LightYield.gamma_cdf

PhotonPropagation.LightYield.get_longitudinal_params

PhotonPropagation.LightYield.integral_long_profile

PhotonPropagation.LightYield.longitudinal_profile

PhotonPropagation.LightYield.rel_additional_track_length

PhotonPropagation.LightYield.rel_additional_track_length_params

PhotonPropagation.LightYield.rescale_source

PhotonPropagation.Medium._absorption_coeff_dust

PhotonPropagation.Medium._absorption_coeff_spice

PhotonPropagation.Medium._absorption_length_straw

PhotonPropagation.Medium._eff_sca_coeff_dust

PhotonPropagation.Medium._ice_group_refractive_index

PhotonPropagation.Medium._ice_phase_refractive_index

PhotonPropagation.Medium.make_homogenous_clearice_properties

PhotonPropagation.PhotonPropagationCuda.initialize_direction_isotropic

PhotonPropagation.PhotonPropagationCuda.uniform

PhotonPropagation.Medium.HomogenousIceProperties — Type

HomogenousIceProperties{T<:Real} <: MediumProperties{T}

Properties for deep (homogenuous) ice. Uses the SPICE model. See: https://user-web.icecube.wisc.edu/~dima/work/WISC/ppc/spice/new/paper/a.pdf

Fields:

-radiationlength – Radiation length (g/cm^2) -meanscatteringangle – Cosine of the mean scattering angle -hgfraction – Mixture weight of the HG component -ASPICE – A parameter of the SPICE model -BSPICE – B parameter of the SPICE model -DSPICE – B parameter of the SPICE model -ESPICE – B parameter of the SPICE model -adust400 – Absorption coefficient at 400nm -bdust400 – Scattering coefficient at 400nm -alphascadust – Alpha parameter of the SPICE model -kappaabsdust – Kappa parameter of the SPICE model -deltaTAMANDA – deltaT parameter of the SPICE model -absscale – Scaling factor for the absorption length -scascale – Scaling factor for the scattering length

PhotonPropagation.Medium._absorption_coeff_dust — Method

_absorption_coeff_dust(wavelength, a_dust_400, kappa_abs_dust)

Calculate the absorption coefficient contribution due to dust Taken from "Measurement of South Pole ice transparency with the IceCube LED calibration system", https://doi.org/10.1016/j.nima.2013.01.054

PhotonPropagation.Medium._absorption_coeff_spice — Method

_absorption_coeff_spice(wavelength, A_SPICE, B_SPICE, a_dust_400, kappa, dT)

Calculate the absorption coefficient for the SPICE model. Taken from "Measurement of South Pole ice transparency with the IceCube LED calibration system", https://doi.org/10.1016/j.nima.2013.01.054

PhotonPropagation.Medium._absorption_length_straw — Method

_absorption_length_straw(wavelength::Real)

Calculate the absorption length at wavelength (in nm).

Based on interpolation of STRAW attenuation length measurement.

PhotonPropagation.Medium._eff_sca_coeff_dust — Method

effscacoeffdust(wavelength, bdust400, alphascadust) Calculate the effective scattering coefficient for the SPICE model. Taken from "Measurement of South Pole ice transparency with the IceCube LED calibration system", https://doi.org/10.1016/j.nima.2013.01.054

PhotonPropagation.Medium._ice_group_refractive_index — Method

icegrouprefractiveindex(wavelength)

Calculate the group refractive index at wavelength (in nm) Taken from: "Role of group and phase velocity in high-energy neutrino observatories", https://doi.org/10.1016/S0927-6505(00)00142-0

PhotonPropagation.Medium._ice_phase_refractive_index — Method

_ice_phase_refractive_index(wavelength)

Calculate the phase refractive index at wavelength (in nm) Taken from: "Role of group and phase velocity in high-energy neutrino observatories", https://doi.org/10.1016/S0927-6505(00)00142-0

PhotonPropagation.Medium.make_homogenous_clearice_properties — Function

make_homogenous_clearice_properties()

Ice properties for the "homogenous clear ice model"

PhotonPropagation.Detection.Circular — Type

Circular{T <: Real} <: TargetShape

Circular target shape. CAUTION: The orientation is currently always e_z

Fields

position - center point of the rectangle (in m)

radius - circle radius (in m)

PhotonPropagation.Detection.DummyShape — Type

DummyShape{T <: Real} <: TargetShape

Dummy target shape, used to define a position.

Fields

position - Shape position

PhotonPropagation.Detection.HomogeneousDetector — Type

struct HomogeneousDetector{TS <: TargetShape} <: PhotonTarget{TS}

A struct representing a homogeneous detector.

Fields

shape::TS: The shape of the detector.

active_area::Float64: The active area of the detector.

module_id::UInt16: The module ID of the detector.

PhotonPropagation.Detection.POMRelativeAcceptance — Type

POMRelativeAcceptance{I} <: PMTAcceptance

This parametrization uses a unit acceptance for PMTs in group 2 and a relative acceptance (relative to group 2) for group 1. This is intended to be used when the emission spectrum is already biased wrt. to the acceptance

PhotonPropagation.Detection.Rectangular — Type

Rectangular{T <: Real} <: TargetShape

Rectangular target shape. CAUTION: The orientation is currently always e_z

Fields

position - center point of the rectangle (in m)

length_x - length in x direction (in m)

length_y - length in y direction (in m)

PhotonPropagation.Detection.Spherical — Type

Spherical{T <: Real} <: TargetShape

Spherical target shape

Fields

position - center point of the sphere (in m)

radius - sphere radius (in m)

PhotonPropagation.Detection.SphericalHomogeneousDetector — Type

struct SphericalHomogeneousDetector{T} <: PhotonTarget{Spherical{T}}

A struct representing a spherical homogeneous detector.

Fields

shape::Spherical{T}: The shape of the detector.

active_area::Float64: The active area of the detector.

module_id::UInt16: The module ID of the detector.

PhotonPropagation.Detection.SphericalMultiPMTDetector — Type

struct SphericalMultiPMTDetector{N,L,T,I} <: PixelatedTarget{Spherical{T}}

A struct representing a spherical multi-PMT detector.

Fields

shape::Spherical{T}: The shape of the detector.

pmt_area::Float64: The area of each PMT.

pmt_coordinates::SMatrix{2,N,Float64,L}: The coordinates of the PMTs.

wl_acceptance::I: The wavelength acceptance of the detector.

module_id::UInt16: The ID of the module.

PhotonPropagation.Detection.TargetShape — Type

TargetShape

Base type for photon target shapes.

PhotonPropagation.Detection.check_pmt_hit — Method

function check_pmt_hit(
    hit_positions::AbstractVector,
    ::AbstractVector,
    ::AbstractVector,
    d::HomogeneousDetector,
    ::Rotation)

Test if photons hit a pmt and return a vector of hit pmt indices (0 for no hit).

Note: This function effectively resamples the (weighted) photons after photon propagation to hits
with their natural rate.

PhotonPropagation.Detection.check_pmt_hit — Method

check_pmt_hit(
    hit_positions::AbstractVector,
    hit_directions::AbstractVector,
    hit_wavelengths::AbstractVector,
    target::POM{<:Real, POMRelativeAcceptance},
    orientation::Rotation{3,<:Real})

Convert photon hits into PMT hits, returning a PMT id for each photon that is detected by a PMT
and `0` for photons that did not hit a PMT.

Here, we only apply the relative wavelength acceptance of PMT group 1 wrt. group 2.
Intended to be used with a biased emission spectrum.

NOTE: DO NOT USE

PhotonPropagation.Detection.check_pmt_hit_opening_angle — Method

check_pmt_hit_opening_angle(rel_hit_position, pmt_positions, opening_angle)

Check if the relative hit position is within the opening angle of any PMT.

Arguments

rel_hit_position::SVector{3,<:Real}: The relative hit position.

pmt_positions::AbstractVector{T}: The positions of the PMTs.

opening_angle::Real: The opening angle threshold.

Returns

j::Int: The index of the first PMT within the opening angle, or 0 if none.

PhotonPropagation.Detection.get_pmt_positions — Method

get_pmt_positions(target, orientation)

Compute the positions of the photomultiplier tubes (PMTs) in a pixelated target.

Arguments

target: A PixelatedTarget object representing the pixelated target.

orientation: A Rotation{3,<:Real} object representing the orientation of the target.

Returns

An array of SVector{3, eltype(target.pmt_coordinates)} representing the positions of the PMTs.

PhotonPropagation.Detection.surface_area — Method

surface_area(::TargetShape) Return the surface area of the target shape.

PhotonPropagation.PhotonPropagationCuda.initialize_direction_isotropic — Method

initialize_direction_isotropic(T::Type)

Sample a direction isotropically

Arguments

T::Type: desired eltype of the return value

Returns

StaticVector{3, T}: Cartesian coordinates of the sampled direction

PhotonPropagation.PhotonPropagationCuda.uniform — Method

uniform(minval::T, maxval::T) where {T}

Convenience function for sampling uniform values in [minval, maxval]

PhotonPropagation.LightYield.AxiconeEmitter — Type

struct AxiconeEmitter{T} <: PhotonSource{T}

A struct representing an axicone emitter.

Fields

position::SVector{3,T}: The position of the emitter in 3D space.

direction::SVector{3,T}: The direction of axis of the axicone.

time::T: The time at which the photons are emitted.

photons::Int64: The number of photons to be emitted.

angle::T: The angle of emission.

beam_divergence::T: The divergence angle of the emitted photons.

PhotonPropagation.LightYield.CherenkovTrackLengthParameters — Type

Struct for storing the parameters of the Cherenkov track length parametrisation. See: https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield.CollimatedIsotropicEmitter — Type

struct CollimatedIsotropicEmitter{T} <: PhotonSource{T}

A struct representing a point-like isotropic photon emitter.

Fields

position::SVector{3,T}: The position of the emitter in 3D space.

time::T: The time at which the photons are emitted.

photons::Int64: The number of photons emitted.

cut_angle_cos: Cosine of the maximum emission angle

PhotonPropagation.LightYield.ExtendedCherenkovEmitter — Type

struct ExtendedCherenkovEmitter{T} <: CherenkovEmitter{T}

A struct representing an extended cascade emitter.

Fields

position::SVector{3,T}: The position of the emitter in three-dimensional space.

direction::SVector{3,T}: The direction of the cascade axis.

time::T: The vertex time of the cascades.

photons::Int64: The number of photons emitted.

long_param::LongitudinalParameterisation{T}: The longitudinal parameterisation of the emitter.

PhotonPropagation.LightYield.ExtendedCherenkovEmitter — Method

ExtendedCherenkovEmitter(particle::Particle, medium::MediumProperties, spectrum::Spectrum; oversample=1.0)

Constructs an ExtendedCherenkovEmitter object that represents the emission of Cherenkov photons by a cascade in a medium.

Arguments

particle::Particle: The particle emitting the Cherenkov photons.

medium::MediumProperties: The properties of the medium.

spectrum::Spectrum: The spectrum of the emitted photons.

oversample::Float64: The oversampling factor for the number of emitted photons (default: 1.0).

Returns

An ExtendedCherenkovEmitter object representing the emission of Cherenkov photons.

PhotonPropagation.LightYield.ExtendedCherenkovEmitterCustomSmear — Type

struct ExtendedCherenkovEmitterMoreSmear{T} <: CherenkovEmitter{T}

A struct representing an extended cascade emitter with additional jitter on the photon emission angle

Fields

position::SVector{3,T}: The position of the emitter in three-dimensional space.

direction::SVector{3,T}: The direction of the cascade axis.

time::T: The vertex time of the cascades.

photons::Int64: The number of photons emitted.

cherenkov_beta_a: "a" parameter of the Beta distribution used to sample Cherenkov track angles.

cherenkov_beta_b: "b" parameter of the Beta distribution used to sample Cherenkov track angles.

long_param::LongitudinalParameterisation{T}: The longitudinal parameterisation of the emitter.

PhotonPropagation.LightYield.ExtendedCherenkovEmitterCustomSmear — Method

ExtendedCherenkovEmitterCustomSmear(particle::Particle, medium::MediumProperties, spectrum::Spectrum; oversample=1.0)

Constructs an ExtendedCherenkovEmitterCustomSmear object that represents the emission of Cherenkov photons by a cascade in a medium. All photons emitted at Cherenkov angle.

Arguments

particle::Particle: The particle emitting the Cherenkov photons.

medium::MediumProperties: The properties of the medium.

spectrum::Spectrum: The spectrum of the emitted photons.

cherenkov_beta_a: "a" parameter of the Beta distribution used to sample Cherenkov track angles.

cherenkov_beta_b: "b" parameter of the Beta distribution used to sample Cherenkov track angles.

oversample::Float64: The oversampling factor for the number of emitted photons (default: 1.0).

Returns

An ExtendedCherenkovEmitterCustomSmear object representing the emission of Cherenkov photons.

PhotonPropagation.LightYield.ExtendedCherenkovEmitterNoSmear — Type

struct ExtendedCherenkovEmitterNoSmear{T} <: CherenkovEmitter{T}

A struct representing an extended cascade emitter where all photons are emitted at Cherenkov angle.

Fields

position::SVector{3,T}: The position of the emitter in three-dimensional space.

direction::SVector{3,T}: The direction of the cascade axis.

time::T: The vertex time of the cascades.

photons::Int64: The number of photons emitted.

long_param::LongitudinalParameterisation{T}: The longitudinal parameterisation of the emitter.

PhotonPropagation.LightYield.ExtendedCherenkovEmitterNoSmear — Method

ExtendedCherenkovEmitterNoSmear(particle::Particle, medium::MediumProperties, spectrum::Spectrum; oversample=1.0)

Constructs an ExtendedCherenkovEmitterNoSmear object that represents the emission of Cherenkov photons by a cascade in a medium. All photons emitted at Cherenkov angle.

Arguments

particle::Particle: The particle emitting the Cherenkov photons.

medium::MediumProperties: The properties of the medium.

spectrum::Spectrum: The spectrum of the emitted photons.

oversample::Float64: The oversampling factor for the number of emitted photons (default: 1.0).

Returns

An ExtendedCherenkovEmitterNoSmear object representing the emission of Cherenkov photons.

PhotonPropagation.LightYield.LightyieldParametrisation — Type

Struct for holding both the longitudinal and the track length parametrisation

PhotonPropagation.LightYield.LongitudinalParameterisation — Type

Struct for storing the energy-corrected values of the longitudinal Cherenkov lightyield parametrisation.

PhotonPropagation.LightYield.LongitudinalParameterisation — Method

LongitudinalParameterisation(energy, medium::MediumProperties, ::Type{ptype})

Construct LongitudinalParameterisation for a particle of type ptype at energy energy

PhotonPropagation.LightYield.LongitudinalParametersBase — Type

LongitudinalParametersBase

Struct for storing the parameters of the longitudinal Cherenkov lightyield parametrisation. These parameters are the baseline values at E=1GeV.

For obtaining the energy-correcected parameters, see get_longitudinal_params.

See: https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield.PencilEmitter — Type

struct PencilEmitter{T} <: PhotonSource{T}

A struct representing a pencil beam that emits photons in a specific direction.

Fields

position::SVector{3,T}: The position of the emitter in 3D space.

direction::SVector{3,T}: The direction in which the photons are emitted.

beam_divergence::T: The divergence angle of the emitted photons.

time::T: The time at which the photons are emitted.

photons::Int64: The number of photons to be emitted.

PhotonPropagation.LightYield.PointlikeIsotropicEmitter — Type

struct PointlikeIsotropicEmitter{T} <: PhotonSource{T}

A struct representing a point-like isotropic photon emitter.

Fields

position::SVector{3,T}: The position of the emitter in 3D space.

time::T: The time at which the photons are emitted.

photons::Int64: The number of photons emitted.

PhotonPropagation.LightYield.PointlikeTimeRangeEmitter — Type

struct PointlikeTimeRangeEmitter{T,U} <: PhotonSource{T}

A struct representing a point-like photon emitter with a time range.

Fields

position::SVector{3,T}: The position of the emitter in 3D space.

time_range::Tuple{U,U}: The range of time in which photons are emitted.

photons::Int64: The number of photons emitted.

PhotonPropagation.LightYield._interp_ch_ang_dist_int — Function

_interp_ch_ang_dist_int(ref_ix_low, ref_ix_hi, steps=0.01)

Interpolate the integral of the Cherenkov angular distance distribution in the range [refixlow, refixhi].

PhotonPropagation.LightYield.cascade_cherenkov_track_length — Method

cascade_cherenkov_track_length(energy::Real, track_len_params::CherenkovTrackLengthParameters)

Cherenkov track length for cascades. energy in GeV

returns track length in m

PhotonPropagation.LightYield.cherenkov_ang_dist — Function

cherenkovangdist(costheta, ref_index)

Angular distribution of cherenkov photons for EM cascades.

Taken from https://arxiv.org/pdf/1210.5140.pdf

PhotonPropagation.LightYield.cherenkov_ang_dist_int_analytic — Function

cherenkov_ang_dist_int_analytic(ref_index, lower, upper, ang_dist_pars)

Integral of the cherenkov angular distribution function.

PhotonPropagation.LightYield.gamma_cdf — Method

gamma_cdf(a, b, z)

Cumulative Gamma distribution $ int_0^z Gamma(a, b) $

PhotonPropagation.LightYield.get_longitudinal_params — Method

get_longitudinal_params(::Type{ptype}, energy)

Calculate the longitudinal parameters at energy (in GeV).

PhotonPropagation.LightYield.integral_long_profile — Method

integral_long_profile(z_low::Real, z_high::Real, long_param::LongitudinalParameterisation)

Integral of the longitudinal profile from shower depth zlow (in m) to zhigh

PhotonPropagation.LightYield.longitudinal_profile — Method

longitudinal_profile(z::Real, medium::MediumProperties, long_param::LongitudinalParameterisation)

Longitudinal shower profile (PDF) at shower depth z ( in m )

PhotonPropagation.LightYield.rel_additional_track_length — Method

rel_additional_track_length(ref_index)

Calculate additional track length for muons. From https://arxiv.org/pdf/1206.5530.pdf

PhotonPropagation.LightYield.rel_additional_track_length_params — Method

rel_additional_track_length_params(ref_index)

Interpolate additional track length parameters, taken from https://arxiv.org/pdf/1206.5530.pdf using E_cut = 500MeV

PhotonPropagation.LightYield.rescale_source — Method

rescale_source(em::T, factor::Number) where {T <: PhotonSource}

Return a new emitter of type T with number of photons increased by factor